+++ tests/Suites/Repository/Generic/AbstractRepositoryTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace AbstractRepo\Test\Suites\Repository\Generic;
+
+use PHPUnit\Framework\TestCase;
+
+class AbstractRepositoryTest extends TestCase
+{
+
+}
